# Copy parsing logic from observe method to observe decorator 