@kitchoi 2020-11-09
> whereas on this branch we get: (notice the times/relative times that are bolded as these are where the improvement was observed - no pun intended)

In the benchmark report, item (1) refers to the time to import the module, i.e. the class construction. That time has increased. In addition to the change in instantiation time, there is a degradation in the import time.

By making instantiation faster, this makes import slower, which could be undesirable in use cases where a project imports a lot of modules (e.g. `from pyface.api import something`) but only uses very few of the imported code.