@aaronayres35 2020-11-09
Note a very similar approach can be taken for the `Property` cases, by adding the same logic to https://github.com/enthought/traits/blob/fa45275d716385bb6ab0af701d919bf1b4230f7c/traits/has_traits.py#L306-L341
before the return statement.
Doing so we see roughly the same 2x speed up instantiation time, but again a slow down in time to import the module/class construction time.  If we decide to continue with this approach, I will open a separate PR for that change as well.  If we decide to try a different approach I will not