@kitchoi 2020-11-09
To parse the expression early in the decorator effectively store the parsed results on the class as a cache.

This was an idea originally floated from @mdickinson: We could also look into applying `functools.lru_cache` on the parse function here:
https://github.com/enthought/traits/blob/fa45275d716385bb6ab0af701d919bf1b4230f7c/traits/observation/parsing.py#L257
This approach might require a little investigation into the memory footprint too (machines nowadays have so much memory... but still there is a limit).

In any case, I would be quite interested in seeing the profiling results to back up these changes, e.g. it'd be good to see a profile which clearly indicates the parse function is the costly step.