@aaronayres35 2020-11-09
> In the benchmark report, item (1) refers to the time to import the module, i.e. the class construction. That time has increased. In addition to the change in instantiation time, there is a degradation in the import time.
> 
> By making instantiation faster, this makes import slower, which could be undesirable in use cases where a project imports a lot of modules (e.g. `from pyface.api import something`) but only uses very few of the imported code.

Theoretically though, for each import we expect >= 1 instantiation as in general you wouldn't expect to import a class if it will never be instantiated.  So an increase in import time, with a comparable decrease in instantiation time would be a net benefit. (correct?)

In this PR if you look at the sum of the case (1) and (2) times you see that between master and this branch the total times are roughly the same (within a second or so) because all this PR really does is move the calls to parse from 1 case to the other.
FWIW I confirmed that on master `parse` is never called in case (1) and is called 10000 times in case (2), and on this branch it is the opposite.