@aaronayres35 2020-11-09
> In any case, I would be quite interested in seeing the profiling results to back up these changes, e.g. it'd be good to see a profile which clearly indicates the parse function is the costly step.

I ran `snakeviz` on the first scenario of having a subclass of HasTraits with a trait defined, e.g. name = Str(), with a method decorated with `@observe("name")` just on case (2) Time to instantiate the HasTraits object. and this is the result (sorted by tottime (the total time spent in the function not including calls to subfunctions)


<img width="1550" alt="Screen Shot 2020-11-09 at 12 02 32 PM" src="https://user-images.githubusercontent.com/36972686/98579177-df755180-2283-11eb-8ea3-11e902b13f47.png">