@kitchoi 2020-11-09
Thank you! It is really good to see the profiles - it confirms parsing being one big chunk of the time spent. We can explore the options to optimize the `_add_or_remove_extra_graphs` code paths too, but that is likely more complicated than just adding a cache for the parsing step. 

I think I have a preference towards `lru_cache` in #1344 for it adds pretty much no complexity to the code structure, at the cost of some memory which would also be needed on this branch (the parsed expressions will be stored on the class).

@mdickinson @aaronayres35 What is your opinion on this?