@aaronayres35 2020-11-09
> Thank you! It is really good to see the profiles - it confirms parsing being one big chunk of the time spent. We can explore the options to optimize the `_add_or_remove_extra_graphs` code paths too, but that is likely more complicated than just adding a cache for the parsing step.
> 
> I think I have a preference towards `lru_cache` in #1344 for it adds pretty much no complexity to the code structure, at the cost of some memory which would also be needed on this branch (the parsed expressions will be stored on the class).
> 
> @mdickinson @aaronayres35 What is your opinion on this?

Yeah I think I agree.  One thing to note is we are only running a benchmark on a few simple cases.  It might be good to try and gauge how well they would both generalize.  For example is a maxsize of 128 appropriate for the lru_cache for most cases?  It seems to work well for the benchmark but I'm not certain that means it will always be sufficient. (I picked it because it is the default and mark suggested in the slack discussion he suspected between 100-200 to be sufficient).