@kitchoi 2020-11-09
I said on this [comment](https://github.com/enthought/traits/pull/1343#issuecomment-724187186):

> it adds pretty much no complexity to the code structure

One complexity to do with caching is mutability: The ObserverExpression objects are treated as immutable, so I think we are good here.

> It might be good to try and gauge how well they would both generalize. For example is a maxsize of 128 appropriate for the lru_cache for most cases? It seems to work well for the benchmark but I'm not certain that means it will always be sufficient.

That's a good point. Perhaps we could try to work out how many ObserverExpression the cache can hold, just to get a sense how realistic the cache is going to be in practice...though that might vary from Python versions to Python versions. Alternatively, the cache size can be set to None (unbound). In any case, I think this might be a discussion that belongs to #1344.