@aaronayres35 2020-11-12
As mentioned in this comment: https://github.com/enthought/traits/pull/1343#discussion_r519885319
I pulled the logic out into a helper function.

Additionally, as mentioned in this comment: https://github.com/enthought/traits/pull/1343#issuecomment-724083538
I used that helper function to address the property case as well